sk-inet: connection in TCP_CLOSE state should not be restored
TODO The control flow in open_inet_sk() is hard to understand, and most
likely buggy.
We should refactor the whole thing to bring some clarity with what paths
are taken depending on the socket state and opts.

We check for tcp_close earlier and skip restore_one_tcp() entirely.

Note in restore_one_tcp():
TCP_LISTEN state cannot happen: the caller checks for
tcp_connect(), which checks for dst_port > 0, which is always 0 in
the TCP_LISTEN state.